requirements.txt
flask==2.3.3
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.1
faiss-cpu==1.8.0
pymysql==1.1.1
nltk==3.9.1
packaging==24.1
cryptography==43.0.1